REFACTOR(Services): reorganize services into domain-based subdirectories.
- Moved existing service files into appropriate subdirectories.

- Modified the namespaces in the relevant files to ensure proper functionality after the restructuring.

- Verified that the services still function correctly after the restructuring.
