+++ backend/app/Http/Controllers/ResetPasswordController.php
@@ -3,7 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\ResetPasswordRequest;
-use App\Services\ResetPasswordService;
+use App\Services\Auth\ResetPasswordService;
 use Illuminate\Http\JsonResponse;
 
 class ResetPasswordController extends Controller
