+++ backend/app/Services/Auth/ForgotPasswordService.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Services;
+namespace App\Services\Auth;
 
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Password;
